backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
aiohttp
requests
beautifulsoup4
python-multipart
numpy
python-dotenv
psycopg2-binary
scikit-learn
pandas
joblib
